requirements.txt
asgiref==3.8.1
Django==4.2.13
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1